Changes states and useffect component
